+++ fetcher/ifeedfetcher.php
@@ -27,12 +27,16 @@ interface IFeedFetcher {
      * no results are fetched
      * @param bool fullTextEnabled if true tells the fetcher to enhance the
      * articles by fetching custom enhanced content
+     * @param string $basicAuthUser if given, basic auth is set for this feed
+     * @param string $basicAuthPassword if given, basic auth is set for this
+     * feed. Ignored if user is null or an empty string
      * @throws FetcherException if the fetcher encounters a problem
      * @return array an array containing the new feed and its items, first
      * element being the Feed and second element being an array of Items
      */
     function fetch($url, $getFavicon=true, $lastModified=null, $etag=null,
-                   $fullTextEnabled=false);
+                   $fullTextEnabled=false, $basicAuthUser=null,
+                   $basicAuthPassword=null);
 
     /**
      * @param string $url the url that should be fetched