+++ js/tests/unit/controller/NavigationControllerSpec.js
@@ -324,7 +324,7 @@ describe('NavigationController', function () {
                         callback({feeds: [{
                             id: 3,
                             url: url,
-                            folderId: folderId
+                            folderId: folderId,
                         }]});
                         return {
                             finally: function (callback) {
@@ -361,7 +361,7 @@ describe('NavigationController', function () {
 
         expect(ctrl.showNewFolder).toBe(false);
         expect(FeedResource.create).toHaveBeenCalledWith('test', 3,
-            undefined);
+            undefined, undefined, undefined);
         expect(Publisher.publishAll).toHaveBeenCalledWith({feeds: [{
             id: 3,
             url: 'test',
@@ -430,14 +430,16 @@ describe('NavigationController', function () {
 
         var feed = {
             url: 'test',
-            newFolder: 'john'
+            newFolder: 'john',
+            user: 'user',
+            password: 'password'
         };
 
         ctrl.createFeed(feed);
 
         expect(ctrl.showNewFolder).toBe(false);
         expect(FeedResource.create).toHaveBeenCalledWith('test', 19,
-            undefined);
+            undefined, 'user', 'password');
         expect(FolderResource.create).toHaveBeenCalledWith('john');
         expect(Publisher.publishAll).toHaveBeenCalledWith({
             folders: [{